Create a config class
* Separation of concerns means logging shouldn't be co-mingled with config.
* Allow non-root docker-compose if local policies allow it.
* If non-root, set up separate per-user directory for configuration
