+++ cmd/maintenanceUpdate/maintenanceUpdate.go
@@ -8,12 +8,12 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/CanastaWiki/Canasta-CLI-Go/internal/canasta"
-	"github.com/CanastaWiki/Canasta-CLI-Go/internal/logging"
+	"github.com/CanastaWiki/Canasta-CLI-Go/internal/config"
 	"github.com/CanastaWiki/Canasta-CLI-Go/internal/orchestrators"
 )
 
 var (
-	instance logging.Installation
+	instance config.Installation
 	pwd      string
 	err      error
 )
@@ -39,7 +39,7 @@ func NewCmdCreate() *cobra.Command {
 	return maintenanceCmd
 }
 
-func runMaintenanceUpdate(instance logging.Installation) {
+func runMaintenanceUpdate(instance config.Installation) {
 	fmt.Println("Running maintenance jobs")
 	orchestrators.Exec(instance.Path, instance.Orchestrator, "web", "php maintenance/update.php && php maintenance/runJobs.php && php canasta-extensions/SemanticMediaWiki/maintenance/rebuildData.php")
 	fmt.Println("Completed running maintenance jobs")
