+++ cmd/restart/restart.go
@@ -7,12 +7,13 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/CanastaWiki/Canasta-CLI-Go/internal/config"
 	"github.com/CanastaWiki/Canasta-CLI-Go/internal/logging"
 	"github.com/CanastaWiki/Canasta-CLI-Go/internal/orchestrators"
 )
 
 func NewCmdCreate() *cobra.Command {
-	var instance logging.Installation
+	var instance config.Installation
 	var verbose bool
 	var restartCmd = &cobra.Command{
 		Use:   "restart",
@@ -42,10 +43,10 @@ func NewCmdCreate() *cobra.Command {
 	return restartCmd
 }
 
-func Restart(instance logging.Installation) error {
+func Restart(instance config.Installation) error {
 	var err error
 	if instance.Id != "" {
-		instance, err = logging.GetDetails(instance.Id)
+		instance, err = config.GetDetails(instance.Id)
 		if err != nil {
 			return err
 		}
