+++ internal/extensionsskins/extensionsskins.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/CanastaWiki/Canasta-CLI-Go/internal/logging"
+	"github.com/CanastaWiki/Canasta-CLI-Go/internal/config"
 	"github.com/CanastaWiki/Canasta-CLI-Go/internal/orchestrators"
 )
 
@@ -23,12 +23,12 @@ func Contains(list []string, element string) bool {
 	return false
 }
 
-func List(instance logging.Installation, constants Item) {
+func List(instance config.Installation, constants Item) {
 	fmt.Printf("Available %s:\n", constants.Name)
 	fmt.Print(orchestrators.Exec(instance.Path, instance.Orchestrator, "web", "cd $MW_HOME/"+constants.RelativeInstallationPath+" && find * -maxdepth 0 -type d"))
 }
 
-func CheckInstalled(name string, instance logging.Installation, constants Item) (string, error) {
+func CheckInstalled(name string, instance config.Installation, constants Item) (string, error) {
 	output := orchestrators.Exec(instance.Path, instance.Orchestrator, "web", "cd $MW_HOME/"+constants.RelativeInstallationPath+" && find * -maxdepth 0 -type d")
 	if !Contains(strings.Split(output, "\n"), name) {
 		return "", fmt.Errorf("%s %s doesn't exist", constants.Name, name)
@@ -36,7 +36,7 @@ func CheckInstalled(name string, instance logging.Installation, constants Item) 
 	return name, nil
 }
 
-func Enable(name string, instance logging.Installation, constants Item) {
+func Enable(name string, instance config.Installation, constants Item) {
 	phpScript := fmt.Sprintf("<?php\n// This file was generated by Canasta\n%s( '%s' );", constants.PhpCommand, name)
 	filePath := fmt.Sprintf("/mediawiki/config/settings/%s.php", name)
 	output, err := orchestrators.ExecWithError(instance.Path, instance.Orchestrator, "web", "ls "+filePath)
@@ -49,7 +49,7 @@ func Enable(name string, instance logging.Installation, constants Item) {
 	}
 }
 
-func CheckEnabled(name string, instance logging.Installation, constants Item) (string, error) {
+func CheckEnabled(name string, instance config.Installation, constants Item) (string, error) {
 	output := orchestrators.Exec(instance.Path, instance.Orchestrator, "web", "ls /mediawiki/config/settings/")
 	if !Contains(strings.Split(output, "\n"), name+".php") {
 		return "", fmt.Errorf("%s %s is not enabled", constants.Name, name)
@@ -61,7 +61,7 @@ func CheckEnabled(name string, instance logging.Installation, constants Item) (s
 	return name, nil
 }
 
-func Disable(name string, instance logging.Installation, constants Item) {
+func Disable(name string, instance config.Installation, constants Item) {
 	command := fmt.Sprintf(`rm /mediawiki/config/settings/%s.php`, name)
 	orchestrators.Exec(instance.Path, instance.Orchestrator, "web", command)
 	fmt.Printf("%s %s disabled\n", constants.Name, name)
